requirements.txt
streamlit
python-dotenv
langchain_community
langchain_openai
langchain_core
langchain-google-genai
